made delete popup modular, created function to get img from AWS
