+++ client/componentUtils/graphPopup.js
@@ -17,3 +17,20 @@ export const GraphPopup = (graph, callback, graphType) => (
         }
     </Popup>
 )
+
+export const DeletePopup = (element, callback, id, keyword) => (
+    <Popup
+    trigger={element}
+    modal
+    closeOnDocumentClick
+  >
+    {close => (
+      <div>
+        <h2>Are you sure you want to delete this {keyword}?</h2>
+        <button onClick={() => callback(id, close)}>Yes, delete</button>
+        <button onClick={close}>No, don't delete</button>
+      </div>
+    )
+    }
+  </Popup>
+)
