+++ server/utils/index.js
@@ -1,0 +1,6 @@
+const { getDatasetFromS3, getGraphImgFromS3 } = require('./AWSutils')
+
+module.exports = {
+    getDatasetFromS3,
+    getGraphImgFromS3
+}
